+++ Source/Kernel/Grains.Specs/EventSequences/for_AppendedEventsQueue/given/a_single_subscriber.cs
@@ -1,0 +1,37 @@
+// Copyright (c) Cratis. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Cratis.Chronicle.Concepts.Events;
+using Cratis.Chronicle.Concepts.Keys;
+using Cratis.Chronicle.Concepts.Observation;
+using Cratis.Chronicle.Grains.Observation;
+
+namespace Cratis.Chronicle.Grains.EventSequences.for_AppendedEventsQueue.given;
+
+public abstract class a_single_subscriber : all_dependencies
+{
+    protected ObserverKey _observerKey;
+    protected IObserver _observer;
+    protected List<HandledEvents> _handledEvents = [];
+    protected AppendedEventsQueue _queue;
+
+    async Task Establish()
+    {
+        _observerKey = new ObserverKey("Some observer", "Some event store", "Some namespace", "Some event sequence");
+        _observer = Substitute.For<IObserver>();
+        _observer
+            .When(_ => _.Handle(Arg.Any<Key>(), Arg.Any<IEnumerable<AppendedEvent>>()))
+            .Do(callInfo =>
+            {
+                var key = callInfo.Arg<Key>();
+                var events = callInfo.Arg<IEnumerable<AppendedEvent>>();
+                _handledEvents.Add(new(key, events));
+            });
+        _grainFactory.GetGrain<IObserver>(_observerKey).Returns(_observer);
+
+        _queue = new AppendedEventsQueue(_taskFactory, _grainFactory);
+        await _queue.Subscribe(_observerKey, EventTypes);
+    }
+
+    protected abstract IEnumerable<EventType> EventTypes { get; }
+}
