+++ Source/Kernel/Grains.Specs/EventSequences/for_AppendedEventsQueue/when_enqueuing/single_event_with_single_subscriber.cs
@@ -5,53 +5,24 @@ using Cratis.Chronicle.Concepts.Events;
 using Cratis.Chronicle.Concepts.Keys;
 using Cratis.Chronicle.Concepts.Observation;
 using Cratis.Chronicle.Grains.Observation;
-using Cratis.Tasks;
 
 namespace Cratis.Chronicle.Grains.EventSequences.for_AppendedEventsQueue.when_enqueuing;
 
-public class single_event_with_single_subscriber : Specification
+public class single_event_with_single_subscriber : given.a_single_subscriber
 {
-    record HandledEvents(Key Partition, IEnumerable<AppendedEvent> Events);
-    ITaskFactory _taskFactory;
-    IGrainFactory _grainFactory;
-
-    AppendedEventsQueue _queue;
-    ObserverKey _observerKey;
-    EventType _eventType;
+    EventType _eventType = new("Some event", 1);
     AppendedEvent _appendedEvent;
-    IObserver _observer;
+
     EventSourceId _eventSourceId;
 
-    List<HandledEvents> _handledEvents = [];
-
-    async Task Establish()
+    void Establish()
     {
-        _taskFactory = Substitute.For<ITaskFactory>();
-        _grainFactory = Substitute.For<IGrainFactory>();
-        _observerKey = new ObserverKey("Some observer", "Some event store", "Some namespace", "Some event sequence");
-        _eventType = new EventType("Some event", 1);
         _appendedEvent = AppendedEvent.EmptyWithEventType(_eventType);
         _eventSourceId = Guid.NewGuid();
         _appendedEvent = _appendedEvent with { Context = EventContext.Empty with { EventSourceId = _eventSourceId } };
-
-        _taskFactory
-            .When(_ => _.Run(Arg.Any<Func<Task>>()))
-            .Do(callInfo => callInfo.Arg<Func<Task>>()());
-
-        _observer = Substitute.For<IObserver>();
-        _observer
-            .When(_ => _.Handle(Arg.Any<Key>(), Arg.Any<IEnumerable<AppendedEvent>>()))
-            .Do(callInfo =>
-            {
-                var key = callInfo.Arg<Key>();
-                var events = callInfo.Arg<IEnumerable<AppendedEvent>>();
-                _handledEvents.Add(new(key, events));
-            });
-        _grainFactory.GetGrain<IObserver>(_observerKey).Returns(_observer);
-
-        _queue = new AppendedEventsQueue(_taskFactory, _grainFactory);
-        await _queue.Subscribe(_observerKey, [_eventType]);
     }
+
+    protected override IEnumerable<EventType> EventTypes => [_eventType];
 
     async Task Because()
     {
@@ -61,4 +32,5 @@ public class single_event_with_single_subscriber : Specification
 
     [Fact] void should_call_handle_on_observer_once() => _handledEvents.Count.ShouldEqual(1);
     [Fact] void should_call_handle_on_observer_with_correct_event_source_id() => _handledEvents[0].Partition.Value.ShouldEqual(_eventSourceId.Value);
+    [Fact] void should_call_handle_on_observer_with_correct_event() => _handledEvents[0].Events.ShouldContainOnly(_appendedEvent);
 }
